ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-chess==1.999
stockfish==3.28.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.15.0
torch==2.1.2
redis==5.0.1
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0
loguru==0.7.2 